Move multiplexer into separate file.
